refactor: 영화 호출 및 리프레시 구조 개편
