+++ src/domain/movieApi.ts
@@ -1,27 +1,12 @@
-import { removeMoreButton } from "../components/MovieList/movieListHandler";
-import { Movie } from "../type";
 import Store from "./Store";
-
-interface MovieResult {
-  poster_path: string;
-  title: string;
-  vote_average: number;
-}
-
-interface MovieApiResponse {
-  page: number;
-  results: MovieResult[];
-  total_pages: number;
-  total_results: number;
-}
 
 const store: Store = Store.getInstance();
 
-export const fetchMovieInfo = async (url: string) => {
+export const fetchMovies = async (url: string) => {
   try {
     const response = await fetch(url).then((data) => data.json());
     if (store.getPage() === response.page) {
-      handleMovieInfoResponse(response);
+      return response;
     }
     else {
       throw new Error(response.status_message);
@@ -32,25 +17,4 @@ export const fetchMovieInfo = async (url: string) => {
   }
 };
 
-export const handleMovieInfoResponse = async (response: MovieApiResponse) => {
-  const { results, total_pages } = await response;
-  store.setTotalPage(total_pages);
-
-  saveMoviesAndRemoveMoreButton(results);
-};
-
-const saveMoviesAndRemoveMoreButton = (results: MovieResult[]) => {
-  store.appendMovies(convertApiResponseToMovieList(results));
-  if (store.getPage() === store.getTotalPage()) removeMoreButton();
-};
-
-const convertApiResponseToMovieList = (results: MovieResult[]): Movie[] => {
-  return results.map((movie) => {
-    return {
-      poster: movie.poster_path,
-      title: movie.title,
-      ratings: movie.vote_average,
-    };
-  });
-};
