+++ src/index.ts
@@ -3,10 +3,10 @@ import "./css/common.css";
 import Header from "./components/Header";
 import MovieList from "./components/MovieList";
 import NoResultsMessage from "./components/NoResultsMessage";
-import { movies } from "./domain/movies";
 import Skeleton from "./components/Skeleton";
+import { updateMovies } from "./domain/movies";
 
-movies.showPopularMovies();
+updateMovies();
 
 customElements.define("movie-header", Header);
 customElements.define("movie-list", MovieList);
